requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
librosa
soundfile
PyYAML